fix: type reference erros
